Propagate exit codes through shell-pipes
Regardless of what the program exists with, the pipes always return 0
which is what scons looks at to determine what to return to the shell.

pipefail sets the $? to the latest non-zero exit code. The parenthesis
spawns a subshell so the setting is contained. pipefail is supported by
(at least) zsh and bash.